Create items for dragzones/dropzones

In Dragula, the drag and drop areas are called "containers". This function creates the individual draggable items and options in these containers.

dragulaZoneItems(zone, type, items, ids = rep(NA, length(items)),
  selected = NULL, invisible = NULL, locked = NULL, freeze = NULL,
  togglevis = FALSE, togglelock = FALSE)

Arguments

zone
Container zone type: either drop or drag

type
Are these options or presets?

items
List of item labels, with names corresponding to values

ids
If multivalued, these will be unique ids

selected
Selected items (array length of items - either NA or ds-selected)

invisible
Invisible items (array length of items - either NA or ds-invisible)

locked
Locked items (array length of items - either NA or ds-locked)

freeze
No items allowed before these. Analogous to freezing the first few columns of a spreadsheet (array length of items - either NA or ds-freeze). Makes since only for first initial items, and when used in conjunction with locked. Frozen items also cannot be toggled.

togglevis
Add an icon to allow toggling items between visible/invisible.

togglelock
Add an icon to allow toggling items between locked/unlocked.

Value

div element